+++ packages/sdks-tests/src/snippet-tests/blueprints-homepage.spec.ts
@@ -12,7 +12,6 @@ test.describe('Homepage', () => {
         'remix',
         'react-sdk-next-15-app',
         'nextjs-sdk-next-app',
-        'angular-19-ssr',
         'gen1-react',
         'gen1-remix',
         'gen1-next14-pages',