+++ packages/sdks/e2e/qwik-city/src/routes/[...index]/index.tsx
@@ -3,6 +3,7 @@ import { routeLoader$ } from '@builder.io/qwik-city';
 import {
   Content,
   _processContentResult,
+  registerAction,
   setClientUserAttributes,
 } from '@builder.io/sdk-qwik';
 import { getProps } from '@sdk/tests';
@@ -91,6 +92,24 @@ export default component$(() => {
       if (window.location.pathname === '/variant-containers/') {
         setClientUserAttributes({
           device: 'tablet',
+        });
+      }
+      if (typeof window !== 'undefined') {
+        registerAction({
+          name: 'test-action',
+          kind: 'function',
+          id: 'test-action-id',
+          inputs: [
+            {
+              name: 'actionName',
+              type: 'string',
+              required: true,
+              helperText: 'Action name',
+            },
+          ],
+          action: () => {
+            return `console.log("function call") `;
+          },
         });
       }
     })