+++ packages/sdks/src/blocks/form/form/form.lite.tsx
@@ -77,6 +77,9 @@ export default function FormComponent(props: FormProps) {
     submissionState(): FormState {
       return (isEditing() && props.previewState) || state.formState;
     },
+    errorResponse(response: any) {
+      return JSON.stringify(response, null, 2);
+    },
     onSubmit(event: any) {
       const sendWithJsProp =
         props.sendWithJs || props.sendSubmissionsTo === 'email';
@@ -103,13 +106,19 @@ export default function FormComponent(props: FormProps) {
           value: File | boolean | number | string | FileList;
         }[] = Array.from(el.querySelectorAll('input,select,textarea'))
           .filter((el) => !!(el as HTMLInputElement).name)
+          .filter(
+            (el) =>
+              !!(el as HTMLInputElement).name &&
+              ((el as HTMLInputElement).type !== 'radio' ||
+                (el as HTMLInputElement).checked)
+          )
           .map((el) => {
             let value: any;
             const key = (el as HTMLImageElement).name;
             if (el instanceof HTMLInputElement) {
               if (el.type === 'radio') {
                 if (el.checked) {
-                  value = el.name;
+                  value = el.value;
                   return { key, value };
                 }
               } else if (el.type === 'checkbox') {
@@ -372,7 +381,7 @@ export default function FormComponent(props: FormProps) {
           class="builder-form-error-text"
           css={{ padding: '10px', color: 'red', textAlign: 'center' }}
         >
-          {JSON.stringify(state.responseData, null, 2)}
+          {state.errorResponse(state.responseData)}
         </pre>
       </Show>