+++ packages/sdks/src/components/awaiter.types.ts
@@ -1,0 +1,7 @@
+export type AwaiterProps = {
+  load: () => Promise<any>;
+  props?: any;
+  attributes?: any;
+  fallback?: any;
+  children?: any;
+};