+++ packages/sdks/output/react-native/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Builder.io React Native SDK Changelog (@builder.io/sdk-react-native)
 
+## 4.2.2
+
+### Patch Changes
+
+- 7adc4f6: Fix: Corrected the implementaion of http requests with GET method
+- 25895a2: Fix: Improved implementation of making Content http-requests with GET method
+
+## 4.2.1
+
+### Patch Changes
+
+- 6ae35cc: feat: add support for register action
+
 ## 4.2.0
 
 ### Minor Changes