+++ packages/sdks/output/react/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Builder.io React SDK Changelog (@builder.io/sdk-react)
 
+## 4.2.2
+
+### Patch Changes
+
+- 7adc4f6: Fix: Corrected the implementaion of http requests with GET method
+- 25895a2: Fix: Improved implementation of making Content http-requests with GET method
+
+## 4.2.1
+
+### Patch Changes
+
+- 6ae35cc: feat: add support for register action
+- 7a0d981: fix: Form submission should use the radio button value rather than name
+
 ## 4.2.0
 
 ### Minor Changes